requirements.txt
streamlit
scikit-learn
pandas
numpy
matplotlib
seaborn
pillow
joblib